feat: add google log in oauth!

+++ project.py
@@ -5,6 +5,18 @@ from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Restaurant, MenuItem
 
+# login session
+from flask import session as login_session
+import random, string
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+from flask import make_response
+import requests
+
+CLIENT_ID = json.loads(
+  open('client_secret.json', 'r').read())['web']['client_id']
 
 #Connect to Database and create database session
 engine = create_engine('sqlite:///restaurantmenu.db')
@@ -13,6 +25,116 @@ Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
+# Log in page
+@app.route('/login')
+def showLogin():
+  state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+    for i in xrange(32))
+  login_session['state'] = state
+  return render_template('login.html', STATE = state)
+
+# Oauth2
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+  # Validate state token
+  if request.args.get('state') != login_session['state']:
+    response = make_response(json.dumps('Invalid state parameter'), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  # Obtain authorization code
+  code = request.data
+  try:
+    # Upgrade the authorization code into a credentials object
+    oauth_flow = flow_from_clientsecrets('client_secret.json', scope='')
+    oauth_flow.redirect_uri = 'postmessage'
+    credentials = oauth_flow.step2_exchange(code)
+  except FlowExchangeError:
+    response = make_response(json.dumps('Failed to upgrade authorization code'), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  # Check that the access token is valid.
+  access_token = credentials.access_token
+  url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s' % access_token)
+  h = httplib2.Http()
+  result = json.loads(h.request(url, 'GET')[1])
+  # If there was an error in the access token info, abort.
+  if result.get('error') is not None:
+    response = make_response(json.dumps(result.get('error')), 500)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  # Verify that the access token is used for the intended user.
+  gplus_id = credentials.id_token['sub']
+  if result['user_id'] != gplus_id:
+    response = make_response(json.dumps("Token's user ID doesn't match given user ID."), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  # Verify that the access token is valid for this app.
+  if result['issued_to'] != CLIENT_ID:
+    response = make_response(json.dumps("Token's client ID doesn't match app's."), 401)
+    print "Token's client ID doesn't match app's."
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  # Check to see if user is already logged in
+  stored_access_token = login_session.get('access_token')
+  stored_gplus_id = login_session.get('gplus_id')
+  if stored_access_token is not None and gplus_id == stored_gplus_id:
+    response = make_response(json.dumps('Current user is already connected.'), 200)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+
+  # Store the access token in the session for later use.
+  login_session['access_token'] = access_token
+  login_session['gplus_id'] = gplus_id
+  # Get user info
+  userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+  params = {'access_token': access_token, 'alt': 'json'}
+  answer = requests.get(userinfo_url, params=params)
+  data = json.loads(answer.text)
+
+  login_session['username'] = data['name']
+  login_session['picture'] = data['picture']
+  login_session['email'] = data['email']
+
+  output = ''
+  output += '<h1>Welcome, '
+  output += login_session['username']
+  output += '!</h1>'
+  output += '<img src="'
+  output += login_session['picture']
+  output += '" style="width: 300px; height: 300px; border-radius: 150px; -webkit-border-radius: 150px; -moz-border-radius: 150px;">'
+  flash("You've logged in as %s" % login_session['username'])
+  return output
+
+# Disconnect oauth2
+@app.route('/gdisconnect')
+def gdisconnect():
+  access_token = login_session.get('access_token')
+  if access_token is None:
+    print 'Access Token is None'
+    response = make_response(json.dumps('Current user not connected.'), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  print 'In gdisconnect access token is %s' % access_token
+  print 'User name is: '
+  print login_session['username']
+  url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+  h = httplib2.Http()
+  result = h.request(url, 'GET')[0]
+  print 'result is '
+  print result
+  if result['status'] == '200':
+    del login_session['access_token']
+    del login_session['gplus_id']
+    del login_session['username']
+    del login_session['email']
+    del login_session['picture']
+    response = make_response(json.dumps('Successfully disconnected.'), 200)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  else:
+    response = make_response(json.dumps('Failed to revoke token for given user.'), 400)
+    response.headers['Content-Type'] = 'application/json'
+    return response
 
 #JSON APIs to view Restaurant Information
 @app.route('/restaurant/<int:restaurant_id>/menu/JSON')
@@ -43,6 +165,8 @@ def showRestaurants():
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
 def newRestaurant():
+  if 'username' not in login_session:
+    return redirect('/login')
   if request.method == 'POST':
       newRestaurant = Restaurant(name = request.form['name'])
       session.add(newRestaurant)
@@ -55,6 +179,8 @@ def newRestaurant():
 #Edit a restaurant
 @app.route('/restaurant/<int:restaurant_id>/edit/', methods = ['GET', 'POST'])
 def editRestaurant(restaurant_id):
+  if 'username' not in login_session:
+    return redirect('/login')
   editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
       if request.form['name']:
@@ -68,6 +194,8 @@ def editRestaurant(restaurant_id):
 #Delete a restaurant
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods = ['GET','POST'])
 def deleteRestaurant(restaurant_id):
+  if 'username' not in login_session:
+    return redirect('/login')
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
     session.delete(restaurantToDelete)
@@ -84,12 +212,14 @@ def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
     return render_template('menu.html', items = items, restaurant = restaurant)
-     
+
 
 
 #Create a new menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/new/',methods=['GET','POST'])
 def newMenuItem(restaurant_id):
+  if 'username' not in login_session:
+    return redirect('/login')
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
       newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
@@ -103,40 +233,41 @@ def newMenuItem(restaurant_id):
 #Edit a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET','POST'])
 def editMenuItem(restaurant_id, menu_id):
-
-    editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
-    restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-    if request.method == 'POST':
-        if request.form['name']:
-            editedItem.name = request.form['name']
-        if request.form['description']:
-            editedItem.description = request.form['description']
-        if request.form['price']:
-            editedItem.price = request.form['price']
-        if request.form['course']:
-            editedItem.course = request.form['course']
-        session.add(editedItem)
-        session.commit() 
-        flash('Menu Item Successfully Edited')
-        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
-    else:
-        return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
+  if 'username' not in login_session:
+    return redirect('/login')
+  editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
+  restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  if request.method == 'POST':
+      if request.form['name']:
+          editedItem.name = request.form['name']
+      if request.form['description']:
+          editedItem.description = request.form['description']
+      if request.form['price']:
+          editedItem.price = request.form['price']
+      if request.form['course']:
+          editedItem.course = request.form['course']
+      session.add(editedItem)
+      session.commit()
+      flash('Menu Item Successfully Edited')
+      return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+  else:
+      return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
 
 
 #Delete a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete', methods = ['GET','POST'])
 def deleteMenuItem(restaurant_id,menu_id):
-    restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-    itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one() 
-    if request.method == 'POST':
-        session.delete(itemToDelete)
-        session.commit()
-        flash('Menu Item Successfully Deleted')
-        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
-    else:
-        return render_template('deleteMenuItem.html', item = itemToDelete)
-
-
+  if 'username' not in login_session:
+    return redirect('/login')
+  restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one()
+  if request.method == 'POST':
+      session.delete(itemToDelete)
+      session.commit()
+      flash('Menu Item Successfully Deleted')
+      return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+  else:
+      return render_template('deleteMenuItem.html', item = itemToDelete)
 
 
 if __name__ == '__main__':
